High Speed Sintering Boundary Example

For an explanation of what exactly this boundary is doing see High Speed Sintering Boundary Explanation

HSSSimulations.HSSBound.HSSParams — Type

An example structure to hold the boundary condition constants used in the example boundary condition functions. T1 must be a function that has the same signature as overheadTempFunc

Fields

name::String

pistonHeat::Any

pistonCool::Any

conductionCoef::Float64

airHeat::Any

airCool::Any

convectionCoef::Float64

surfaceHeat::Any

surfaceCool::Any

powderTemp::Any

sinterLamp::Vector{Float64}

recoatLamp::Vector{Float64}

lampWidth::Int64

lampOffset::Int64

carriageWidth::Int64

printCarriageWidth::Int64

printOffset::Int64

recoatOffset::Int64

surfaceTarget::Float64

surfaceTol::Float64

overheadHeatupFunc::Any

overheadLayerStep::Int64

overheadPowerStep::Float64

overheadMaxPower::Float64

percentOverhead::Float64

overheadPower::Float64: Current power input of overhead heaters

overheadTemp::Float64: Current temperature of overhead heaters

lastUpdatedOverhead::Int64: The last layer the overheads were updated on

coolStart::Float64: The time at the start of cooling loads

pistonCoolStart::Float64: The offset time if the piston wasn't at max temp at start of cooling

airCoolStart::Float64: The offset time if the air wasn't at max temp at start of cooling

surfaceCoolStart::Float64: The offset time if the machine's internal suraface wasn't at max temp at start of cooling

HSSSimulations.HSSBound.HSSParams — Method

HSSParams(
    geometry::Geometry;
    name,
    pistonPath,
    airPath,
    surfacePath,
    conductionCoef,
    lampVector,
    lampWidth,
    lampOffset,
    carriageWidth,
    recoatOffset,
    printCarriageWidth,
    printOffset,
    surfaceTarget,
    surfaceTol,
    overheadLayerStep,
    overheadPercentStep,
    overheadTemp,
    overheadPower,
    overheadPowerOut,
    overheadHeatCapacity,
    overheadMaxPower,
    convectionCoef,
    sinterPower,
    recoatPower,
    lastUpdatedOverhead,
    percentOverhead,
    powderTempDelta,
    overheadHeatupFunc
) -> HSSParams{_A, _B, _C, _D, _E, _F, HSSSimulations.HSSBound.var"#16#21"{Geometry, HSSSimulations.HSSBound.var"#15#20", Float64}} where {_A, _B, _C, _D, _E, _F}

The geometry (of type Geometry) should be the same one used for the simulation. If the piston target temperature is chaneged then the piston path will need to be changed to curves that will match the target temperature. Same if the preheat bed is thicker than the normal ≈3 mm.

HSSSimulations.HSSBound.HSSLoads — Function

HSSLoads(
    skip,
    geometry;
    nrPreheat,
    lenPreheat,
    nrCool,
    lenCool,
    sinterSpeed,
    lcAndBedWidth
) -> Vector{AbstractLoadSet}

Returns a list of loads for the preheat, build and cooldown loadsets for a default build for the HSS example. The same skip is used for all load steps (See Why We Skip Some Results for more information on skip).

julia> HSSLoads(10, Geometry((1,1,1),1,1); nrPreheat=5, lenPreheat=60.0, nrCool=5, lenCool=60.0, sinterSpeed=0.160)
3-element Vector{AbstractLoadSet}:
   name : Preheat
  loads
----------------------

  Name: Overheads Only
  For 5 loads

   name : Layer
  finishLayer : 0
  loads
----------------------

  Name: Overheads Only
  Name: Sintering
  Name: Overheads Only
  Name: Recoating
  Name: Overheads Only
  Name: No Inking
  Name: Overheads Only
  Name: Inking

   name : Cooldown
  loads
----------------------

  Name: Overheads Off
  For 5 loads

Loads

HSSSimulations.HSSBound.loadCooldown — Function

loadCooldown(
    tₗ,
    skip
) -> Load{SymetryBoundary, SymetryBoundary, SymetryBoundary, SymetryBoundary}

Uses HSSBound.OverheadsCoolBoundary and HSSBound.PistonCoolBoundary so that the overhead and piston heaters are turned off. Assumes that once cooling starts it doesn't stop.

HSSSimulations.HSSBound.loadSinterStroke — Function

loadSinterStroke(
    tₗ,
    skip
) -> Load{SymetryBoundary, SymetryBoundary, SymetryBoundary, SymetryBoundary}

Returns the compleate Types.Load struct for a HSS build with recoat/lamp carriage over the build and the recoat lamp set to sinter power. This assumes that a subset of the build is being simulated and the the edge boundaries can be approximated as symetrical (no heat flow).

Uses HSSBound.SinterBoundary for the top surface, HSSBound.PistonBoundary for the bottom surface and the default Boundary.SymetryBoundary for the sides.

HSSSimulations.HSSBound.loadRecoatStroke — Function

loadRecoatStroke(
    tₗ,
    skip
) -> Load{SymetryBoundary, SymetryBoundary, SymetryBoundary, SymetryBoundary}

Returns the compleate Types.Load struct for a HSS build with the recoat/lamp carriage over the build area, the lamp set to recoat power and new powder being deposited movement. This assumes that a subset of the build is being simulated and the the edge boundaries can be approximated as symetrical (no heat flow).

Uses HSSBound.RecoatBoundary for the top surface, HSSBound.PistonBoundary for the bottom surface and the default Boundary.SymetryBoundary functions for the sides.

HSSSimulations.HSSBound.loadInkStroke — Function

loadInkStroke(
    tₗ,
    skip
) -> Load{SymetryBoundary, SymetryBoundary, SymetryBoundary, SymetryBoundary}

Returns the compleate Types.Load struct for a HSS build with the print head carriage over the build volume depositing ink. This assumes that a subset of the build is being simulated and the the edge boundaries can be approximated as symetrical (no heat flow).

Uses HSSBound.InkBoundary for the top surface, HSSBound.PistonBoundary for the bottom surface and the default Boundary.SymetryBoundary for the sides.

HSSSimulations.HSSBound.loadBlankStroke — Function

loadBlankStroke(
    tₗ,
    skip
) -> Load{SymetryBoundary, SymetryBoundary, SymetryBoundary, SymetryBoundary}

Returns the compleate Types.Load struct for a HSS build with the print head carriage over the build area but not on the ink deposition stroke. This assumes that a subset of the build is being simulated and the the edge boundaries can be approximated as symetrical (no heat flow).

Uses HSSBound.BlankBoundary for the top surface, HSSBound.PistonBoundary for the bottom surface and the default Boundary.SymetryBoundary functions for the sides.

Utilities

HSSSimulations.HSSBound.lampMaker — Function

lampMaker(
    lampVector::Vector,
    lampWidth,
    geometry::Geometry
) -> Any

Returns a vector of heat flux density coefficients representing the lamp, with the same node spacing as the simulation. The returned vector can be multiplied by the lamp power, in watts, to get a vector of heat flux densities of the lamp.

Arguments

lampVector::Vector : A vector represeting the heat distribution of the lamp in the y axis

lampWidth : The total width represented by the lamp vector (in the y axis, in nodes).

geometry::Geometry : The simulation geometry

Note

If the lamp width divided by the length of the lamp vector is not equal to the Δy then linear interpolation is used to fill in the values.

Warn

This uses the Δx and Δy node spacings to calculate the area used to convert to heat flux density, because of this it is only applicable to the z₁ and z₂ boundaries.

HSSSimulations.HSSBound.overheadTempFunc — Function

overheadTempFunc(
    powerIn,
    powerOut,
    overheadHeatCapacity,
    Δt,
    prevOverheadTemp
) -> Any

Calculates the new temperature of a an overhead heater that has a given power output and is set to a given power (powerIn) values.

Arguments

powerOut : A function that takes the temperatu of the heater and returns the power output

overheadHeatCapacity : The heat capacity of the overhead heater

prevOverheadTemp : The previous temperature of the overhead heaters

HSSSimulations.HSSBound.coolingStart — Function

coolingStart(tₚ, t, params::AbstractProblemParams)

Sets the time that the cooling starts, to be used to calculate how far into the cooling the simulation is during future time steps.

It also finds how far into the cooling curves of the different components to start based on the current temperature of that component. If the current temperature is less than any temperature in the cooling curve it will default to starting at the end of the curve.

This requires the type of pistonCool, airCool and surfaceHeat have methods for findfirst