Material Examples

PA2200

Provided is an example material property structs based on EOS's PA2200 powder. Calling it with no arguments will return the relevant MatProp struct.

The sources of the data used are summarized below. For more details, check the material model chapter of my thesis.

HSSSimulations.Material.PA_κ — Function

Values from Yaun 2014 used for consolidation state of 0 and 1 (at a melt state of 0). And values from Riedlbaur 2015 (I think originally from Rietzel 2011, but my German isn't good enough to confirm) used for a melt state of 1.

HSSSimulations.Material.PA_ρ — Function

Assumes density is constant across temperatures and at data sheet values for maximum consolidation state. For consolidation and melt state zero, the measured value of the powder in the machine is used. The melted density is based on observations while taking dilatometry measurments.

HSSSimulations.Material.PA_c — Function

Data taken from stepscan of PA2200, processed by perkin-elmer dsc software. Melt region is interpolated over to avoide issuse with .

HSSSimulations.Material.PA_Mᵣ — Function

A normalized cumulative integral of the energy in to a system during melting from a DSC curves.

HSSSimulations.Material.PA_Rᵣ — Function

A normalized cumulative integral of the energy out of a system during recrystallization from DSC curves. With each curve taken from increasing melt states.

HSSSimulations.Material.PA_Hf — Function

The difference between the energy in to the system during melting (from a DSC curve), and an interpolated line drawn between the sections of heating outside of the melt region.

HSSSimulations.Material.PA_Hr — Function

The difference between the energy out of the system during crystallization (from a DSC curve), and an interpolated line drawn between the sections of cooling outside of the recrystallization region.

HSSSimulations.Material.PA_Ċ — Function

Based on model from Childs and Tontowi 2001. Modified to work off of consolidation and melt states. Calibrated from dilatometry data. See Material.PA_Ċ_maker for how to make your own.

HSSSimulations.Material.PA_eₚ — Constant

Taken from near ir experiments.

HSSSimulations.Material.PA_eᵢ — Constant

Taken from near ir experiments.

HSSSimulations.Material.PA_ε — Constant

Taken from pyrometer calibrations.