Advanced API

These APIs are exported by their module, but not exported by the package, so to use them you will need to either use their full address:

julia> using HSSSimulations
julia> Solver.loadSetSolver!loadSetSolver! (generic function with 2 methods)

or by using the module it is in:

julia> using HSSSimulations
julia> using .Solver
julia> loadSetSolver!loadSetSolver! (generic function with 2 methods)

Custom Load Sets

HSSSimulations.Solver.loadSetSolver! — Function

loadSetSolver!(
    loadSet<:AbstractLoadSet
    initResult::AbstractResult,
    layerNum::Int,
    prob::Problem{T,Gh,Mp,R,OR,B},
) where {T<:Any,Gh<:Any,Mp<:Any,R<:Any,OR<:Any,B<:Any}

This function is run once for each LoadSet passed in to the problem, and dispatched on the type of that load set. For basic load sets, like FixedLoadSet, it should basically just pass the arguments on to the innerLoadSetSolver! function, the one exception being the list of loads being passed in instead of the AbstractLoadSet struct. For more complicated load sets, like LayerLoadSet, this should include the logic used for that load set. For LayerLoadSet this is a loop that loops over the list of loads once for each layer, incrementing the layer number as it goes.

For examples see the source code for the two default implementations.

HSSSimulations.Types.AbstractLoadSet — Type

abstract type AbstractLoadSet <: HSSSimulations.Types.AbstractSimProperty

This struct holds the information for a load set, primarily the list of loads, but also some extra bits. Subtypes should implement a method for the Solver.loadSetSolver! function.

HSSSimulations.Solver.FixedLoadSet — Type

struct FixedLoadSet <: AbstractLoadSet

#Fields

name::String: Name of the load set

loads::Vector{Load}: List of loads for load set

HSSSimulations.Solver.loadSetSolver! — Method

loadSetSolver!(
    loadSet::FixedLoadSet,
    initResult::AbstractResult,
    layerNum::Int64,
    prob::Problem{T, Gh, Mp, R, OR, B}
) -> Tuple{AbstractResult, Int64}

HSSSimulations.Solver.LayerLoadSet — Type

struct LayerLoadSet <: AbstractLoadSet

#Fields

name::String: Name of the load set

finishLayer::Int64: The last layer to deposit as part of this load set

loads::Vector{Load}: List of loads for load set

HSSSimulations.Solver.loadSetSolver! — Method

loadSetSolver!(
    loadSet::LayerLoadSet,
    initResult::AbstractResult,
    initLayerNum::Int64,
    prob::Problem{T, Gh, Mp, R, OR, B}
) -> Tuple{AbstractResult, Int64}

HSSSimulations.Solver.innerLoadSetSolver! — Function

innerLoadSetSolver!(
    loads::Vector{Load},
    initResult::AbstractResult,
    layerNum,
    prob::Problem;
    name,
    recoat,
    prevLayerNum
)

Solves a load set of a problem, iterating over and solving the loads in the given load set in order. The results for a load is saved to the file after solving it.

If the list of loads passed given include powder deposition, the recoat kwarg should be set to true, with the layerNum set to the thickness of the powder bed in layers that is desired at the end of the loads. If this is more than one layer thicker than the previous value, then the previous values should also be passed in to prevLayerNum.

It also helps to provide a unique name. This is the name that will be used to store the results from runnning the list of loads in the output file, and also will be used as the label in the progress meter. If no name is given then the current time is used instead. If the name given is not unique then you will get a run time error when it tries to save results to a place that is already taken.

Custom Materials

HSSSimulations.Material.MatProp — Type

struct MatProp{T1, T2, T3, T4, T5, T6, T7, T8, T9} <: AbstractMatProp

The default implementation of AbstractMatProp used by the default Material.calcMatProps! method. See Material Examples for some premade materials. The arguments to its constructor are the same as the matching fields. The only addition is the geometry argument, which should just be the Geometry struct of for the simulation.

Warning

The maximum melt state is never reset, so if a node goes from the melt region to the recrystalisation region more than once in a simulation, the behaviour might not be modeled properly.

Fields

ρ::Any: Density. Two dimensional interpolations, with inputs in axis 1 = consolidation state, axis 2 = melt state.

c::Any: Specific heat capacity. One dimensional interpolations with the node's temperature as it's input.

κ::Any: Thermal conductivity. Three dimensional interpolations, with inputs in axis 1 = consolidation state, axis 2 = temperature and axis 3 = melt state.

Mᵣ::Any: Melting range and temp relationship. A 1d interpolation, with the only input axis being the temperature and the output is the melt state.

Mₛ::Float64: The start temperature of the melt range

Mₑ::Float64: The end temperature of the melt range

Rᵣ::Any: Crystallisation range and temp relationship. A 2d interpolation, with the first input axis being the node temperature and the second input axis being the maximum melt state reached. The output is the melt state.

Rₛ::Float64: The start temperature of the recrystalisation range

Rₑ::Float64: The end temperature of the recrystalisation range

Hf::Any: Total specific heat of melting

Hr::Any: Total specific heat of fusion. A 1d Interpolation where the input is the maximum melt state reached and the output is the total heat of fusion.

Ċ::Any: Consolidation rate. Either a single value representing a constant consolidation rate, or a 2d interpolation, with the first axis being the node temperature and the second being the node melt state. The output is the rate of change of sinter state.

eₚ::Float64: Emmisivity of the powder. This is the emmisivity relative to the lamp. So the emmisivity over the range of the wavelengths that the lamp outputs, scaled by the relative output power of the lamp at those wavelengths.

eᵢ::Float64: Emmisivity of the powder with ink on it. This is the emmisivity relative to the lamp. So the emmisivity over the range of the wavelengths that the lamp outputs, scaled by the relative output power of the lamp at those wavelengths.

ε::Float64: Black body emmisivity of the powder

name::String: Just used for future reference

Mₘ::Any: Used by the material model to track the maximum melt state reached per node

HSSSimulations.Material.MatProp — Method

MatProp(
    ρ,
    c,
    κ,
    Mᵣ,
    Rᵣ,
    Hf,
    Hr,
    Ċ,
    eₚ,
    eᵢ,
    ε,
    name,
    geometry
) -> MatProp

Default constructor for MatProp.

Custom Material Models

HSSSimulations.Types.AbstractMatProp — Type

abstract type AbstractMatProp <: HSSSimulations.Types.AbstractSimProperty

The abstract type that any material property struct needs to be a subtype of. These are for storing the values used to define the properties of a material, and those used to calculate the changes in material properties due to the build. The default struct used is MatProp, however new subtypes can be defined and used by defining a new Material.calcMatProps! function. This allows for the simplificiation (or complication) of the material model used in the simulation to fit your needs.

Any new subtype will need it's own method writen for calcMatProps! and if the example boundaries are being used some of them might need to be updated to use the new material properties. If melt state or consolidation state are beind removed from the material model then it might be worth implementing a new AbstractResult to save on memory. If an additional property that needs to be tracked is added to the material model then it will be necissary to implement a new subtype of those to track this.

See Material Model Recipes and Tutorial 3: A Melt Rate Based Material Model for more information on implementing your own.

HSSSimulations.Material.calcMatProps! — Function

calcMatProps!(::AbstractResult, ::AbstractResult, <:Problem, ::Indices)

Updates any of the material properties that change, along with the Fourier number for the current time step. New methods can be defined by dispatching on Problem with a different subtype of MatProp (for more information on this, see Types.AbstractMatProp). The two main functions called by this one, meltUpdate and consUpdate, are both also dispatched on the type of the material property struct, so they can be overwritten individually for a new material property struct if the new material model only needs to change some of the behaviour.

Arguments

cts: Current time step results

pts: Previous time step results

ind: The indicies to update

HSSSimulations.Material.calcMatProps! — Method

calcMatProps!(
    pts::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp<:AbstractMatProp, R, OR, B},
    ind
)

The default material model, designed to be used with MatProp, but should work with any Types.AbstractMatProp that shares MatProp's ρ, c, κ and Mₘ fields.

Updates melt and consolidation state, and then uses those to update the density, heat capacity and conductivity. These are then used to calculate the Fourier number for each axis for this time step.

HSSSimulations.Material.meltUpdate — Function

meltUpdate(Mᵗ⁻¹, T, Mₘ, Δt, mp::AbstractMatProp) -> Mᵗ, Mₘ, Δh

Calculates the new melt state of a single node. It is given the melt state and temperature of the node in the previous time step, the maximum melt state the node has reached so far, the duration of the current time step and the material property struct.

It should return the new melt state for the node, the new maximum melt state, and the specific enthalpy change due to fusion or recrystalisation that is associated with the change in melt state.

HSSSimulations.Material.meltUpdate — Method

meltUpdate(
    Mᵗ⁻¹,
    T,
    Mₘ,
    _,
    mp::AbstractMatProp
) -> Tuple{Any, Any, Any}

Updates the melt state of a node if the temperature of the node is in the melting or crystalisation termperature range if not the existing value is returned. This uses a basic time independent model that associates a single melt state to a given combination of temperature and maximum melt state reached so far.

HSSSimulations.Material.consUpdate — Function

consUpdate(C, M, T, Δt, mp::AbstractMatProp) -> C

Calculates the new consolidation state of a single node. It is given the current consolidation state, melt state and temperature of the node, the duration of the current time step and the material property struct.

HSSSimulations.Material.consUpdate — Method

consUpdate(C, M, T, Δt, mp::AbstractMatProp) -> Any

Updates the consolidation state of a node by adding the change in consolidation state (rate * time) to the previous consolidation state. Only applies if the material is melted and not already fully consolidated.

This finds the consolidation rate by calling mp.Ċ with C, T and M as arguments.

The returned values maximum is limited to 1.

HSSSimulations.Material.PA_Ċ_maker — Function

Based on model from Childs and Tontowi 2001. Modified to work off of consolidation and melt states. Takes three coefficients and returns a consolidation rate calculator.

Custom Results

HSSSimulations.Types.AbstractResult — Type

abstract type AbstractResult <: HSSSimulations.Types.AbstractSimProperty

This as the abstract type for the results from a single simulation time step. See Tutorial 4: Saving More Results and Result Type Recipes for how to make your own subtype.

Warning

As this struct is what stores the data during the simulation, all subtypes MUST have the T, t and tₚ fields, and if used with the default material model it will also need the M and C fields.

HSSSimulations.Res.loadStepSaver — Function

loadSave(loadResultsFolder, loadResults::StructVector{T}) where {T<:AbstractResult}

Saves the results of a load step into the folder for the current load step in the output file of the problem, given by the loadResultsFolder argument.

Any new methods for this function should dispatch on the type parameter of the loadResults, and save any desired results to the file like so:

function loadStepSaver(loadResultsFolder, loadResults::StructVector{T}) where {T<:AbstractResult}
    loadResultsFolder["time"] = loadResults.t
    loadResultsFolder["T"] = stack(loadResults.T)
    return
end

Note

stack is used as a very efficient way of converting the vector of 3D arrays from the StructVector in to a 4D array.

HSSSimulations.Res.loadStepSaver — Method

loadStepSaver(
    loadResultsFolder,
    loadResults::StructArray{T<:Result, 1}
)

Saves the results for the temperature, melt state and consolidation state as 4D arrays of X,Y,Z and t. And the time is saved as a 1D Array.

HSSSimulations.Types.AbstractOtherResults — Type

abstract type AbstractOtherResults <: HSSSimulations.Types.AbstractSimProperty

This as the abstract type for the results that are saved only at the end of the simulation. This is useful to store things that do not change every time step, without requiring any changes to the material property or params structs. However those structs can also be used to store such things, such as is done for Mₘ in the MatProp struct.

As the default implementation does nothing, and nothing normally dispatches on it, it can be replaced by another empty type to use a new method for any function that dispatches on Problem (I think that's all of the user facing API).

Tutorial 4: Saving More Results for more details.

HSSSimulations.Types.OtherResults — Type

struct OtherResults <: AbstractOtherResults

The default struct stores no data and only acts as a placeholder.

HSSSimulations.Res.otherResults — Function

otherResults(prob<:Problem, file)

Runs at the end of the simulation to save any additional results that only need to be saved once, as opposed to for every nth time step. The Types.Problem type and any of its type parameters can be dispatched on.

See AbstractOtherResults for a place to store random data, and Tutorial 4: Saving More Results for a detailed guide.

HSSSimulations.Res.otherResults — Method

otherResults(problem::Problem{T, Gh, Mp, R, OR, B}, file)

Default otherResults method that just saves the maximum melt state reached for each node.

Custom Boundaries

HSSSimulations.Types.AbstractProblemParams — Type

abstract type AbstractProblemParams <: HSSSimulations.Types.AbstractSimProperty

The Problem struct contains a field called boundary that is of this type. The purpose of this field is to store any boundary condition inputs that changes over the course of a build, such as the air temperature in the machine. As this is an abstract type a new struct can be made, that is a subtype of AbstractProblemParams, to contain the required variables for a given simulation. An example of this is given in HSSParams.

Any new subtype will need it's own boundaries written (the examples given are only meant to work with the HSSParams example).

See the Problem Parameters Recipes for some insperation for implementing your own.

HSSSimulations.Types.boundaryHeatTransferRate — Function

boundaryHeatTransferRate(T, i, p<:AbstractBoundary) -> ϕ⃗::Float64

Used to calculate the heat flux denstity (ϕ⃗, in wm⁻²) into the model at a node on the relavant face for a given boundary condition. They are run per node, for each node on the face. The function should return a single numeric value that represents the heat flux density into the model.

Arguments

T : The temperature of the node for which the ghost node is being calculated

i : The cartesian index of the node for which the ghost node is being calculated

p : The boundary struct for the current boundary (see AbstractBoundary)

See the implementations of this for examples, and Boundary Recipes for some additional examples along with the relevant boundary structs.

Note

A positive heat flow equates to heat flowing into the model and negative out of the model.

Warning

As this is run for every node on the face every time step, try to avoid complex computations. Where possible, put them in the constructor of the AbstractBoundary instead, as that is only run once per face per timestep.

HSSSimulations.Types.boundaryHeatTransferRate — Method

boundaryHeatTransferRate(
    _,
    _,
    _::SymetryBoundary
) -> Float64

HSSSimulations.Types.boundaryHeatTransferRate — Method

boundaryHeatTransferRate(T, _, p::ConductionBoundary) -> Any

HSSSimulations.Types.boundaryHeatTransferRate — Method

boundaryHeatTransferRate(T, _, p::ConvectionBoundary) -> Any

HSSSimulations.Types.boundaryHeatTransferRate — Method

boundaryHeatTransferRate(
    T,
    _,
    p::HSSSimulations.HSSBound.AbstractPistonBoundary
) -> Any

HSSSimulations.Types.boundaryHeatTransferRate — Method

boundaryHeatTransferRate(
    T,
    _,
    p::HSSSimulations.HSSBound.AbstractOverheadsBoundary
) -> Any

HSSSimulations.Types.boundaryHeatTransferRate — Method

boundaryHeatTransferRate(T, i, p::SinterBoundary) -> Any

HSSSimulations.Types.boundaryHeatTransferRate — Method

boundaryHeatTransferRate(T, i, p::RecoatBoundary) -> Any

HSSSimulations.Types.boundaryHeatTransferRate — Method

boundaryHeatTransferRate(T, i, p::InkBoundary) -> Any

HSSSimulations.Types.boundaryHeatTransferRate — Method

boundaryHeatTransferRate(T, i, p::BlankBoundary) -> Any

HSSSimulations.Types.AbstractBoundary — Type

Subtypes of this define a boundary for a single face of the model. An instance of the subtype is created per boundary each time step.

The fields of the subtype should contain all of the information needed to calculate the heatflow on that boundary (except for the information already passed in to Types.boundaryHeatTransferRate).

Each subtype of this should have a constructor that satisfies the following signature:

<Boundary>(pts::AbstractResult, cts::AbstractResult, prob::Problem, ls::LoadStep)

Arguments

pts : The results from the previous timestep. See Result

cts : The results from the current timestep. See Result

prob : The 'global' variables of the simulation. See Problem

ls : The per load step variables. See Types.LoadStep

For examples see the default subtypes, and their constructors. Boundary Recipes has examples of some of the extra things that can be done with new AbstractBoundarys.

Warning

The constructor for z₂ is the only one that is run before the indices lists are set for that load. So any updating of the indices, such as that done by recoating!, should be done in that boundary constructor.

Warning

cts.T is unknown when this function is called, use the temperature from the previous time step insted (pts.T).

HSSSimulations.Boundary.conductionFlow — Function

conductionFlow(T1, T2, h)

Calculates a heat flux density (ϕ⃗, in wm⁻²) for a given contact conduction coefficient, h, between points at temperature T1 (the node on the boundary) and T2 (the wall in contact with the node).

HSSSimulations.Boundary.convectionFlow — Function

convectionFlow(T1, T∞, h)

Calculates a heat flux density (ϕ⃗, in wm⁻²) for a given convection coefficient, h, between a point at temperature T1 and a fluid with temperature T∞.

HSSSimulations.Boundary.radiationFlow — Function

radiationFlow(T1, T∞, ε)

Calculates a heat flux density (ϕ⃗, in wm⁻²) for a given emmisivity, ε, between a point at temperature T1 and infinity at temperature T∞. The temperature arguments should both be in the units °C.

HSSSimulations.HSSBound.AbstractPistonBoundary — Type

abstract type AbstractPistonBoundary <: AbstractBoundary

HSSSimulations.HSSBound.PistonBoundary — Type

struct PistonBoundary <: HSSSimulations.HSSBound.AbstractPistonBoundary

Boundary for the bottom surface of a HSS build for loads where the heater is turned on. Assumes that the piston maintains a constant temperature at after the initial heat up.

Fields

piston::Float64: Piston temperature

h::Float64: Contact conduction coefficient

HSSSimulations.HSSBound.PistonBoundary — Method

PistonBoundary(
    _::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    _::HSSSimulations.Types.LoadStep
) -> PistonBoundary

HSSSimulations.HSSBound.PistonCoolBoundary — Type

struct PistonCoolBoundary <: HSSSimulations.HSSBound.AbstractPistonBoundary

Boundary for the bottom surface of a HSS build for cooldown loads. Assumes that once cooling starts it doesn't stop.

Fields

piston::Float64: Piston temperature

h::Float64: Contact conduction coefficient

HSSSimulations.HSSBound.PistonCoolBoundary — Method

PistonCoolBoundary(
    pts::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    _::HSSSimulations.Types.LoadStep
) -> PistonCoolBoundary

HSSSimulations.HSSBound.AbstractOverheadsBoundary — Type

abstract type AbstractOverheadsBoundary <: AbstractBoundary

HSSSimulations.HSSBound.OverheadsBoundary — Type

struct OverheadsBoundary <: HSSSimulations.HSSBound.AbstractOverheadsBoundary

Boundary for the top surface of a HSS build with no lamp or print carriage movement. Assumes the heater controller is the same as our HSS machine where It only updates the overhead power every overheadLayerStep number of layers, and only once on that layer.

Fields

overheadTemp::Float64: Temperatur of overhead heater

surfaceTemp::Float64: Temperatur of machine internal surface

ε::Float64: Black Body Emmissivity

airTemp::Float64: Temperature of air above build bed

h::Float64: Convection coefficient

Po::Float64: Weighting of overhead heaters vs raditiation loss

HSSSimulations.HSSBound.OverheadsBoundary — Method

OverheadsBoundary(
    pts::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    ls::HSSSimulations.Types.LoadStep
) -> OverheadsBoundary

HSSSimulations.HSSBound.OverheadsCoolBoundary — Type

struct OverheadsCoolBoundary <: HSSSimulations.HSSBound.AbstractOverheadsBoundary

Like the OverheadsBoundary but it sets the overhead power to 0w (turning it off). Assumes that once cooling starts it doesn't stop.

Fields

overheadTemp::Float64: Temperatur of overhead heater

surfaceTemp::Float64: Temperatur of machine internal surface

ε::Float64: Black Body Emmissivity

airTemp::Float64: Temperature of air above build bed

h::Float64: Convection coefficient

Po::Float64: Weighting of overhead heaters vs raditiation loss

HSSSimulations.HSSBound.OverheadsCoolBoundary — Method

OverheadsCoolBoundary(
    pts::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    _::HSSSimulations.Types.LoadStep
) -> OverheadsCoolBoundary

HSSSimulations.HSSBound.SinterBoundary — Type

struct SinterBoundary <: AbstractBoundary

Boundary for the top surface of a HSS build for when the recoat/lamp carriage is over the build bed and the lamp is set to sinter power.

Fields

overheadTemp::Float64: Temperatur of overhead heater

surfaceTemp::Float64: Temperatur of machine internal surface

eₗ::Array{Float64, 3}: Net absorbtivity relative to lamp

ε::Float64: Black Body Emmissivity

airTemp::Float64: Temperature of air above build bed

h::Float64: Convection coefficient

lamp::Vector{Float64}: Sinter lamp array

shadow::Vector{Bool}: Carriage shadow array

Po::Float64: Weighting of overhead heaters vs raditiation loss

HSSSimulations.HSSBound.SinterBoundary — Method

SinterBoundary(
    _::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    ls::HSSSimulations.Types.LoadStep
) -> SinterBoundary

HSSSimulations.HSSBound.RecoatBoundary — Type

struct RecoatBoundary <: AbstractBoundary

Boundary for the top surface of a HSS build with the recoat/lamp carriage over the build area, the lamp set to recoat power and new powder being deposited.

Fields

overheadTemp::Float64: Temperatur of overhead heater

surfaceTemp::Float64: Temperatur of machine internal surface

eₗ::Array{Float64, 3}: Net absorbtivity relative to lamp

ε::Float64: Black Body Emmissivity

airTemp::Float64: Temperature of air above build bed

h::Float64: Convection coefficient

lamp::Vector{Float64}: Sinter lamp array

shadow::Vector{Bool}: Carriage shadow array

Po::Float64: Weighting of overhead heaters vs raditiation loss

HSSSimulations.HSSBound.RecoatBoundary — Method

RecoatBoundary(
    pts::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    ls::HSSSimulations.Types.LoadStep
) -> RecoatBoundary

HSSSimulations.HSSBound.InkBoundary — Type

Boundary for the top surface of a HSS build with the print head carriage over the build volume depositing ink.

Fields

overheadTemp::Float64: Temperatur of overhead heater

surfaceTemp::Float64: Temperatur of machine internal surface

ε::Float64: Black Body Emmissivity

airTemp::Float64: Temperature of air above build bed

h::Float64: Convection coefficient

shadow::Vector{Bool}: Carriage shadow array

Po::Float64: Weighting of overhead heaters vs raditiation loss

HSSSimulations.HSSBound.InkBoundary — Method

InkBoundary(
    _::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    ls::HSSSimulations.Types.LoadStep
) -> InkBoundary

HSSSimulations.HSSBound.BlankBoundary — Type

struct BlankBoundary <: AbstractBoundary

Boundary for the top surface of a HSS build with the print head carriage over the build area but not on the ink deposition stroke.

Fields

overheadTemp::Float64: Temperatur of overhead heater

surfaceTemp::Float64: Temperatur of machine internal surface

ε::Float64: Black Body Emmissivity

airTemp::Float64: Temperature of air above build bed

h::Float64: Convection coefficient

shadow::Vector{Bool}: Carriage shadow array

Po::Float64: Weighting of overhead heaters vs raditiation loss

HSSSimulations.HSSBound.BlankBoundary — Method

BlankBoundary(
    _::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    ls::HSSSimulations.Types.LoadStep
) -> BlankBoundary

HSSSimulations.HSSBound.loadOverheads — Function

loadNoTop(tₗ, skip)

Returns the compleate HSSBound.Load struct for a HSS build with no lamp or print carriage movement. This assumes that a subset of the build is being simulated and the the edge boundaries can be approximated as symetrical (no heat flow).

Uses HSSBound.OverheadsBoundary for the top surface, HSSBound.PistonBoundary for the bottom surface and the default Boundary.SymetryBoundary functions for the sides.

HSSSimulations.Types.Load — Method

Types.Load(;
    name="default", tₗ=1, skip=1,
    x₁=SymetryBoundary, x₂=SymetryBoundary,
    y₁=SymetryBoundary, y₂=SymetryBoundary,
    z₁=SymetryBoundary, z₂=SymetryBoundary,
)

Default constructor for Types.Load. All of the boundaries defaults to symetry boundaries (see Boundary.SymetryBoundary). See Boundary.ConductionBoundary and Boundary.ConvectionBoundary, for some more built in loads. And High Speed Sintering Boundary Example for more complicated loads.

The z₂ load is run before any others, so any calculations that need to be done first should be put there, such as recoating.

Examples

julia> Load(; name="test", tₗ=1, skip=1, x₁=ConductionBoundary, y₂=ConvectionBoundary)
  x₁ : ConductionBoundary
  x₂ : SymetryBoundary
  y₁ : SymetryBoundary
  y₂ : ConvectionBoundary
  z₁ : SymetryBoundary
  z₂ : SymetryBoundary
  name : test
  tₗ : 1.0
  skip : 1

HSSSimulations.Types.Load — Type

struct Load{X₁<:AbstractBoundary, X₂<:AbstractBoundary, Y₁<:AbstractBoundary, Y₂<:AbstractBoundary, Z₁<:AbstractBoundary, Z₂<:AbstractBoundary} <: HSSSimulations.Types.AbstractSimProperty

Fields

x₁::Type{X₁} where X₁<:AbstractBoundary: The boundary condition for the start of the x axis ([1,:,:])

x₂::Type{X₂} where X₂<:AbstractBoundary: The boundary condition for the end of the x axis ([end,:,:])

y₁::Type{Y₁} where Y₁<:AbstractBoundary: The boundary condition for the start of the y axis ([:,1,:])

y₂::Type{Y₂} where Y₂<:AbstractBoundary: The boundary condition for the end of the y axis ([:,end,:])

z₁::Type{Z₁} where Z₁<:AbstractBoundary: The boundary condition for the start of the z axis ([:,:,1])

z₂::Type{Z₂} where Z₂<:AbstractBoundary: The boundary condition for the end of the z axis ([:,:,end])

name::String: Used to annotate the results

tₗ::Float64: The duration of the load in seconds

skip::Int64: The distance between results to save (see Why We Skip Some Results)

Each of the boundaries given must satisfy Boundary isa Type{T} where T <: AbstractBoundary. This is to say that they should be the type itself, not an instance of that type. Additionally it should be a subtype of AbstractBoundary and meet all of the requirements outlined in the that documentation.

Note

If the Load is in a BuildLoadSet then one of the boundaries in one of the loads in the set should call the recoating logic so that new powder is actuall being layed down. This should always be done in the z₂ boundary, as it is a special case that ensures the indicies are updated for all other boundaries.

Basic Boundaries

HSSSimulations.Boundary.SymetryBoundary — Type

struct SymetryBoundary <: AbstractBoundary

A boundary for a face that has no heatflow (such as a face on a symetrical boundary). Used as the default boundary in Types.Load.

HSSSimulations.Boundary.ConductionBoundary — Type

struct ConductionBoundary <: AbstractBoundary

A boundary for a face that has a source of contact conductive heat flow.

Fields

temp::Float64: Temperature of object being contacted

coef::Float64: Coefficient of contact conduction

HSSSimulations.Boundary.ConvectionBoundary — Type

struct ConvectionBoundary <: AbstractBoundary

A boundary for a face that has a source of convective heat flow.

Fields

temp::Float64: Air temperature

coef::Float64: Convection coefficient

Utilities

HSSSimulations.Boundary.movingObjOverlap — Function

movingObjOverlap(
    geometry::Geometry,
    movingObj,
    objPos::Tuple{Int64, Int64}
) -> Vector

Makes a vector the length of the y-axis of the simulation. It represents the overlap between the simulation area an object that moves in the y-axis, whos left and right edge are given by objPos. The overlap is filled with the value passed as movingObj.

The difference between the first and second value of the objPos Tuple multipled by the spacing in the y-axis (Δy) should match the width of the moving object.

Examples

julia> geometry = Geometry((0.015, 0.015, 0.010), 0.005, 0.010; offset=(0, 0.01));

julia> movingObjOverlap(geometry, true, (2, 4))
3-element Vector{Bool}:
 1
 1
 0

movingObjOverlap(
    geometry::Geometry,
    movingObj::AbstractVector,
    objPos::Int64
) -> Any

Makes a vector the length of the y-axis of the simulation. It represents the overlap between the simulation area an object that moves in the y-axis, whos right edge are given by objPos. The overlap is filled with the matching values from the vector movingObj.

The length of the vector multipled by the spacing in the y-axis (Δy) should match the width of the moving object.

Examples

julia> geometry = Geometry((0.015, 0.015, 0.010), 0.005, 0.010; offset=(0, 0.01));

julia> movingObjOverlap(geometry, [1, 2, 3, 4, 5], 6)
3-element Vector{Int64}:
 2
 3
 4

julia> movingObjOverlap(geometry, [1, 2, 3, 4, 5], 8)
3-element Vector{Int64}:
 0
 1
 2

movingObjOverlap(
    geometry::Geometry,
    movingObj::AbstractMatrix,
    objPos::Int64
) -> Any

Makes a matrix with the same dimensions as the top surface of the simulation area. It represents the overlap between the simulation area an object that moves in the y-axis, whos right edge are given by objPos. The overlap is filled with the matching values from the matrix movingObj.

Also for Matrix inputs, the length of the first axis multipled by the spacing in the x-axis (Δx) should match the depth of the moving object and the length of the second axis multipled by the spacing in the y-axis (Δy) should match the width of the moving object.

Examples

julia> geometry = Geometry((0.015, 0.015, 0.010), 0.005, 0.010; offset=(0, 0.01));

julia> movingObjOverlap(geometry, [1 2 3; 4 5 6; 7 8 9], 4)
3×3 Matrix{Int64}:
 2  3  0
 5  6  0
 8  9  0

HSSSimulations.Boundary.recoating! — Function

recoating!(
    pts::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    ls::HSSSimulations.Types.LoadStep,
    recoatDist,
    temp
)

Powder recoating logic. This updates the indices and fills the new real indices with the initial results provided in the simulation's Problem, except the temperature, which is set to temp. This allows for the new powder temp to be set to the temperature of the powder in the hopper, or feed bed, at that time in the build. This always runs in the positive y axis, so new powder is always deposited starting from [:, 1, :] and going to [:, end, :].

Logging

HSSSimulations.Solver.makeLogger — Function

makeLogger(debugGroups, logFile) -> LoggingExtras.TeeLogger

debugGroups is a list of groups of debug messages to log to the logFile. If it is set to true then all log groups in the package are used (see package_groups), if it is set to false then no log groups are used.

User log groups can be added to the list, simple add your own string to the list and then set debug messages with the _group field set to your string.

There is a special group called "misc" that will catch any group not in the package_groups list or in any additional group given by the user to the debugGroups list.