HSS Boundary

HSSSimulations.HSSBound — Module

All additional boundary conditions requried to simulate a HSS build that are not already included in Boundary.

For an explanation of what exactly this boundary is doing see High Speed Sintering Boundary Explanation

Parametres

HSSSimulations.HSSBound.HSSParams — Type

An example structure to hold the boundary condition constants used in the example boundary condition functions. T1 must be a function that has the same signature as overheadTempFunc

Fields

name::String

pistonHeat::Any

pistonCool::Any

conductionCoef::Float64

airHeat::Any

airCool::Any

convectionCoef::Float64

surfaceHeat::Any

surfaceCool::Any

powderTemp::Any

sinterLamp::Vector{Float64}

recoatLamp::Vector{Float64}

lampWidth::Int64

lampOffset::Int64

carriageWidth::Int64

printCarriageWidth::Int64

printOffset::Int64

recoatOffset::Int64

surfaceTarget::Float64

surfaceTol::Float64

overheadHeatupFunc::Any

overheadLayerStep::Int64

overheadPowerStep::Float64

overheadMaxPower::Float64

percentOverhead::Float64

overheadPower::Float64: Current power input of overhead heaters

overheadTemp::Float64: Current temperature of overhead heaters

lastUpdatedOverhead::Int64: The last layer the overheads were updated on

coolStart::Float64: The time at the start of cooling loads

pistonCoolStart::Float64: The offset time if the piston wasn't at max temp at start of cooling

airCoolStart::Float64: The offset time if the air wasn't at max temp at start of cooling

surfaceCoolStart::Float64: The offset time if the machine's internal suraface wasn't at max temp at start of cooling

No Carriage Loads

HSSSimulations.HSSBound.loadCooldown — Function

loadCooldown(
    tₗ,
    skip
) -> Load{SymetryBoundary, SymetryBoundary, SymetryBoundary, SymetryBoundary}

Uses HSSBound.OverheadsCoolBoundary and HSSBound.PistonCoolBoundary so that the overhead and piston heaters are turned off. Assumes that once cooling starts it doesn't stop.

HSSSimulations.HSSBound.loadOverheads — Function

loadNoTop(tₗ, skip)

Returns the compleate HSSBound.Load struct for a HSS build with no lamp or print carriage movement. This assumes that a subset of the build is being simulated and the the edge boundaries can be approximated as symetrical (no heat flow).

Uses HSSBound.OverheadsBoundary for the top surface, HSSBound.PistonBoundary for the bottom surface and the default Boundary.SymetryBoundary functions for the sides.

Piston Boundaries

HSSSimulations.Types.boundaryHeatTransferRate — Method

boundaryHeatTransferRate(
    T,
    _,
    p::HSSSimulations.HSSBound.AbstractPistonBoundary
) -> Any

HSSSimulations.HSSBound.AbstractPistonBoundary — Type

abstract type AbstractPistonBoundary <: AbstractBoundary

HSSSimulations.HSSBound.PistonBoundary — Type

struct PistonBoundary <: HSSSimulations.HSSBound.AbstractPistonBoundary

Boundary for the bottom surface of a HSS build for loads where the heater is turned on. Assumes that the piston maintains a constant temperature at after the initial heat up.

Fields

piston::Float64: Piston temperature

h::Float64: Contact conduction coefficient

HSSSimulations.HSSBound.PistonBoundary — Method

PistonBoundary(
    _::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    _::HSSSimulations.Types.LoadStep
) -> PistonBoundary

HSSSimulations.HSSBound.PistonCoolBoundary — Type

struct PistonCoolBoundary <: HSSSimulations.HSSBound.AbstractPistonBoundary

Boundary for the bottom surface of a HSS build for cooldown loads. Assumes that once cooling starts it doesn't stop.

Fields

piston::Float64: Piston temperature

h::Float64: Contact conduction coefficient

HSSSimulations.HSSBound.PistonCoolBoundary — Method

PistonCoolBoundary(
    pts::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    _::HSSSimulations.Types.LoadStep
) -> PistonCoolBoundary

Overhead Boundaries

HSSSimulations.Types.boundaryHeatTransferRate — Method

boundaryHeatTransferRate(
    T,
    _,
    p::HSSSimulations.HSSBound.AbstractOverheadsBoundary
) -> Any

HSSSimulations.HSSBound.AbstractOverheadsBoundary — Type

abstract type AbstractOverheadsBoundary <: AbstractBoundary

HSSSimulations.HSSBound.OverheadsBoundary — Type

struct OverheadsBoundary <: HSSSimulations.HSSBound.AbstractOverheadsBoundary

Boundary for the top surface of a HSS build with no lamp or print carriage movement. Assumes the heater controller is the same as our HSS machine where It only updates the overhead power every overheadLayerStep number of layers, and only once on that layer.

Fields

overheadTemp::Float64: Temperatur of overhead heater

surfaceTemp::Float64: Temperatur of machine internal surface

ε::Float64: Black Body Emmissivity

airTemp::Float64: Temperature of air above build bed

h::Float64: Convection coefficient

Po::Float64: Weighting of overhead heaters vs raditiation loss

HSSSimulations.HSSBound.OverheadsBoundary — Method

OverheadsBoundary(
    pts::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    ls::HSSSimulations.Types.LoadStep
) -> OverheadsBoundary

HSSSimulations.HSSBound.OverheadsCoolBoundary — Type

struct OverheadsCoolBoundary <: HSSSimulations.HSSBound.AbstractOverheadsBoundary

Like the OverheadsBoundary but it sets the overhead power to 0w (turning it off). Assumes that once cooling starts it doesn't stop.

Fields

overheadTemp::Float64: Temperatur of overhead heater

surfaceTemp::Float64: Temperatur of machine internal surface

ε::Float64: Black Body Emmissivity

airTemp::Float64: Temperature of air above build bed

h::Float64: Convection coefficient

Po::Float64: Weighting of overhead heaters vs raditiation loss

HSSSimulations.HSSBound.OverheadsCoolBoundary — Method

OverheadsCoolBoundary(
    pts::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    _::HSSSimulations.Types.LoadStep
) -> OverheadsCoolBoundary

Sinter Stroke

HSSSimulations.HSSBound.loadSinterStroke — Function

loadSinterStroke(
    tₗ,
    skip
) -> Load{SymetryBoundary, SymetryBoundary, SymetryBoundary, SymetryBoundary}

Returns the compleate Types.Load struct for a HSS build with recoat/lamp carriage over the build and the recoat lamp set to sinter power. This assumes that a subset of the build is being simulated and the the edge boundaries can be approximated as symetrical (no heat flow).

Uses HSSBound.SinterBoundary for the top surface, HSSBound.PistonBoundary for the bottom surface and the default Boundary.SymetryBoundary for the sides.

HSSSimulations.Types.boundaryHeatTransferRate — Method

boundaryHeatTransferRate(T, i, p::SinterBoundary) -> Any

HSSSimulations.HSSBound.SinterBoundary — Type

struct SinterBoundary <: AbstractBoundary

Boundary for the top surface of a HSS build for when the recoat/lamp carriage is over the build bed and the lamp is set to sinter power.

Fields

overheadTemp::Float64: Temperatur of overhead heater

surfaceTemp::Float64: Temperatur of machine internal surface

eₗ::Array{Float64, 3}: Net absorbtivity relative to lamp

ε::Float64: Black Body Emmissivity

airTemp::Float64: Temperature of air above build bed

h::Float64: Convection coefficient

lamp::Vector{Float64}: Sinter lamp array

shadow::Vector{Bool}: Carriage shadow array

Po::Float64: Weighting of overhead heaters vs raditiation loss

HSSSimulations.HSSBound.SinterBoundary — Method

SinterBoundary(
    _::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    ls::HSSSimulations.Types.LoadStep
) -> SinterBoundary

Recoating Stroke

HSSSimulations.HSSBound.loadRecoatStroke — Function

loadRecoatStroke(
    tₗ,
    skip
) -> Load{SymetryBoundary, SymetryBoundary, SymetryBoundary, SymetryBoundary}

Returns the compleate Types.Load struct for a HSS build with the recoat/lamp carriage over the build area, the lamp set to recoat power and new powder being deposited movement. This assumes that a subset of the build is being simulated and the the edge boundaries can be approximated as symetrical (no heat flow).

Uses HSSBound.RecoatBoundary for the top surface, HSSBound.PistonBoundary for the bottom surface and the default Boundary.SymetryBoundary functions for the sides.

HSSSimulations.Types.boundaryHeatTransferRate — Method

boundaryHeatTransferRate(T, i, p::RecoatBoundary) -> Any

HSSSimulations.HSSBound.RecoatBoundary — Type

struct RecoatBoundary <: AbstractBoundary

Boundary for the top surface of a HSS build with the recoat/lamp carriage over the build area, the lamp set to recoat power and new powder being deposited.

Fields

overheadTemp::Float64: Temperatur of overhead heater

surfaceTemp::Float64: Temperatur of machine internal surface

eₗ::Array{Float64, 3}: Net absorbtivity relative to lamp

ε::Float64: Black Body Emmissivity

airTemp::Float64: Temperature of air above build bed

h::Float64: Convection coefficient

lamp::Vector{Float64}: Sinter lamp array

shadow::Vector{Bool}: Carriage shadow array

Po::Float64: Weighting of overhead heaters vs raditiation loss

HSSSimulations.HSSBound.RecoatBoundary — Method

RecoatBoundary(
    pts::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    ls::HSSSimulations.Types.LoadStep
) -> RecoatBoundary

Printing Stroke

HSSSimulations.HSSBound.loadInkStroke — Function

loadInkStroke(
    tₗ,
    skip
) -> Load{SymetryBoundary, SymetryBoundary, SymetryBoundary, SymetryBoundary}

Returns the compleate Types.Load struct for a HSS build with the print head carriage over the build volume depositing ink. This assumes that a subset of the build is being simulated and the the edge boundaries can be approximated as symetrical (no heat flow).

Uses HSSBound.InkBoundary for the top surface, HSSBound.PistonBoundary for the bottom surface and the default Boundary.SymetryBoundary for the sides.

HSSSimulations.Types.boundaryHeatTransferRate — Method

boundaryHeatTransferRate(T, i, p::InkBoundary) -> Any

HSSSimulations.HSSBound.InkBoundary — Type

Boundary for the top surface of a HSS build with the print head carriage over the build volume depositing ink.

Fields

overheadTemp::Float64: Temperatur of overhead heater

surfaceTemp::Float64: Temperatur of machine internal surface

ε::Float64: Black Body Emmissivity

airTemp::Float64: Temperature of air above build bed

h::Float64: Convection coefficient

shadow::Vector{Bool}: Carriage shadow array

Po::Float64: Weighting of overhead heaters vs raditiation loss

HSSSimulations.HSSBound.InkBoundary — Method

InkBoundary(
    _::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    ls::HSSSimulations.Types.LoadStep
) -> InkBoundary

Printhead Shadow Only Stroke

HSSSimulations.HSSBound.loadBlankStroke — Function

loadBlankStroke(
    tₗ,
    skip
) -> Load{SymetryBoundary, SymetryBoundary, SymetryBoundary, SymetryBoundary}

Returns the compleate Types.Load struct for a HSS build with the print head carriage over the build area but not on the ink deposition stroke. This assumes that a subset of the build is being simulated and the the edge boundaries can be approximated as symetrical (no heat flow).

Uses HSSBound.BlankBoundary for the top surface, HSSBound.PistonBoundary for the bottom surface and the default Boundary.SymetryBoundary functions for the sides.

HSSSimulations.Types.boundaryHeatTransferRate — Method

boundaryHeatTransferRate(T, i, p::BlankBoundary) -> Any

HSSSimulations.HSSBound.BlankBoundary — Type

struct BlankBoundary <: AbstractBoundary

Boundary for the top surface of a HSS build with the print head carriage over the build area but not on the ink deposition stroke.

Fields

overheadTemp::Float64: Temperatur of overhead heater

surfaceTemp::Float64: Temperatur of machine internal surface

ε::Float64: Black Body Emmissivity

airTemp::Float64: Temperature of air above build bed

h::Float64: Convection coefficient

shadow::Vector{Bool}: Carriage shadow array

Po::Float64: Weighting of overhead heaters vs raditiation loss

HSSSimulations.HSSBound.BlankBoundary — Method

BlankBoundary(
    _::AbstractResult,
    cts::AbstractResult,
    prob::Problem{T, Gh, Mp, R, OR, B},
    ls::HSSSimulations.Types.LoadStep
) -> BlankBoundary

Utilities

HSSSimulations.HSSBound.lampMaker — Function

lampMaker(
    lampVector::Vector,
    lampWidth,
    geometry::Geometry
) -> Any

Returns a vector of heat flux density coefficients representing the lamp, with the same node spacing as the simulation. The returned vector can be multiplied by the lamp power, in watts, to get a vector of heat flux densities of the lamp.

Arguments

lampVector::Vector : A vector represeting the heat distribution of the lamp in the y axis

lampWidth : The total width represented by the lamp vector (in the y axis, in nodes).

geometry::Geometry : The simulation geometry

Note

If the lamp width divided by the length of the lamp vector is not equal to the Δy then linear interpolation is used to fill in the values.

Warn

This uses the Δx and Δy node spacings to calculate the area used to convert to heat flux density, because of this it is only applicable to the z₁ and z₂ boundaries.

HSSSimulations.HSSBound.overheadTempFunc — Function

overheadTempFunc(
    powerIn,
    powerOut,
    overheadHeatCapacity,
    Δt,
    prevOverheadTemp
) -> Any

Calculates the new temperature of a an overhead heater that has a given power output and is set to a given power (powerIn) values.

Arguments

powerOut : A function that takes the temperatu of the heater and returns the power output

overheadHeatCapacity : The heat capacity of the overhead heater

prevOverheadTemp : The previous temperature of the overhead heaters

HSSSimulations.HSSBound.coolingStart — Function

coolingStart(tₚ, t, params::AbstractProblemParams)

Sets the time that the cooling starts, to be used to calculate how far into the cooling the simulation is during future time steps.

It also finds how far into the cooling curves of the different components to start based on the current temperature of that component. If the current temperature is less than any temperature in the cooling curve it will default to starting at the end of the curve.

This requires the type of pistonCool, airCool and surfaceHeat have methods for findfirst