Internal APIs

These are things exported from modules, but not intended for external use.

Types

HSSSimulations.Types.AbstractSimProperty — Type

abstract type AbstractSimProperty

This is the parent type of all of the simulation property types. It exists mostly to define a show method for all of it's subtype. This just makes things nicer to look at the.

HSSSimulations.Types.GVars — Type

struct GVars{T<:AbstractArray, Gh<:AbstractArray, M<:AbstractMatProp, R<:AbstractResult, OR<:AbstractOtherResults, P<:AbstractProblemParams}

An internal struct used for passing the many variables around the different function, it is built from a given problem struct, using a constructor in the problem module. There is probably a better way of doing this, but it works.

Fields

geometry::Geometry: Simulation geometry

matProp::AbstractMatProp: Simulation material properties

params::AbstractProblemParams: Parameters used for boundary conditions

Fx::AbstractArray: Matrix of x axis Fourier numbers

Fy::AbstractArray: Matrix of y axis Fourier numbers

Fz::AbstractArray: Matrix of z axis Fourier numbers

κ::AbstractArray: Matrix of thermal conductivities

Tᵗ⁻¹::AbstractArray: Ghost node padded matrix of previous time step

eᵗ::AbstractArray: Matrix of current emissivities

ink::Ink: Matrix of final emissivities

file::String: File name

init::AbstractResult: Initial results

otherResults::AbstractOtherResults: Results struct used to save data only once at the end

resultsIndex::Vector{String}: index of results file

options::Options: Simulation options

HSSSimulations.Types.GVars — Method

GVars(
    problem::Problem
) -> GVars{Array{Float64, 3}, OffsetArrays.OffsetArray{Float64, 3, Array{Float64, 3}}}

Make an instance of Types.GVars from a Problem.

HSSSimulations.Types.LoadTime — Type

struct LoadTime

Contains all of the time information for a LoadStep.

Fields

Δt::Float64: Time between time steps

times::Vector{Float64}: Array of all of the times in the load step

unskipTimes::Vector{Float64}: Array of all none skipped times in the load step

tₚ::Vector{Float64}: Array of percentage through timestep, same index as times

utₚ::Vector{Float64}: Array of percentage through timestep, same index as unskipTimes

tₛ::Float64: The time at the start of the load step

tₑ::Float64: Time step end time

HSSSimulations.Types.LoadTime — Method

LoadTime(tₛ, tₗ, Δt, skip) -> LoadTime

Default Constructor for LoadTime.

Arguments

tₛ::Float64: The time at the start of the load step

tₗ: The lenght of the load step

Δt::Float64: Time between time steps

skip: How often to save results. See Why We Skip Some Results

HSSSimulations.Types.LoadStep — Type

struct LoadStep{R<:AbstractResult}

An internal struct for the propeties that are constant within a load step.

Fields

time::LoadTime: See LoadTime

size::Tuple{Int64, Int64, Int64}: The x, y and z size in number of nodes (mostly just used for the z, which may have changed from the same value in the GVars.geometry variable)

ind::Indices: See Indices

load::Load: See Load

init::AbstractResult: The initial results for the load. See AbstractResult

name::String: Used for the progress meter

layerNum::Int64: Used for any Load functions that need the layer number. For preheat and cooldown load sets this will be the initial thickness or the finial thickness in layers respectively.

HSSSimulations.Types.Indices — Type

mutable struct Indices

Stores the indices of the nodes, taking into account if the node represents a volume that contains powder or not. And also stores the node pais for the boundaries.

iᵣ are indices of real nodes, this is all nodes that represent space where there is currently powder. This will normally include every node, however before and during the recoat load it will not include all of the top layer, as powder has not been deposited there yet. iᵢ are the indices of 'imaginary' nodes, these are nodes that represent locations that do not currently contain powder. These is the same as the indices of all of the nodes minus the indicies of the real nodes (iᵣ).

Fields with a subscript 1 (₁) denote the boundary at the start of that axis, and subscript 2 (₂) dentoes the end of that boundary. eg. x₂ is the boundary [end,:,:] and y₁ is y[:,1,:]. z₂ is a special case where it always represents the top surface of the build, even if a layer has not been compleatly deposited yet. y₂ is another special case, where it represents its normal face, but also represents the leading edge of deposited powder.

The first element of the tuple in the boundary indices is the index (of the array without the ghost cells) for the real node and the second element is the index (of the array with the ghost cells) of the matching ghost node.

Fields

iᵣ::Vector{CartesianIndex{3}}: List of currently real nodes

iᵢ::Vector{CartesianIndex{3}}: List of currently imaginary nodes

x₁::Matrix{Tuple{CartesianIndex{3}, CartesianIndex{3}}}

x₂::Matrix{Tuple{CartesianIndex{3}, CartesianIndex{3}}}

y₁::Matrix{Tuple{CartesianIndex{3}, CartesianIndex{3}}}

y₂::Matrix{Tuple{CartesianIndex{3}, CartesianIndex{3}}}: List of real/ghost node pairs for `[:,end,:]

z₁::Matrix{Tuple{CartesianIndex{3}, CartesianIndex{3}}}

z₂::Matrix{Tuple{CartesianIndex{3}, CartesianIndex{3}}}: List of real/ghost node pairs for `[:,:,end]

iₘ::CartesianIndex{3}: The middle point of z₂, and it is currently only used for debugging/logging.

ΔH::Int64: The same as it is in Geometry, it is just copied over for convenience.

HSSSimulations.Types.Geometry — Type

struct Geometry <: HSSSimulations.Types.AbstractSimProperty

Defines the geometry of the build volume (and the simulation volume within that (if only a subset of a build is being simulated).

Fields

x::Float64: The size of the x axis of the model in meters

y::Float64: The size of the y axis of the model in meters

z::Float64: The size of the z axis of the model in meters

X::Int64: The size of the x axis of the model in nodes

Y::Int64: The size of the y axis of the model in nodes

Z::Int64: The size of the z axis of the model in nodes

xoffset::Float64: The x axis offset of the model from the machines datum in meters

yoffset::Float64: The y axis offset of the model from the machines datum in meters

X_OFFSET::Int64: The x axis offset of the model from the machines datum in nodes

Y_OFFSET::Int64: The y axis offset of the model from the machines datum in nodes

xbuild::Float64: The total x axis size of the machine from which the subset is taken in meters

ybuild::Float64: The total y axis size of the machine from which the subset is taken in meters

X_BUILD::Int64: The total x axis size of the machine from which the subset is taken in nodes

Y_BUILD::Int64: The total y axis size of the machine from which the subset is taken in nodes

Δx::Float64: The spacing of the nodes in meters in the x axis

Δy::Float64: The spacing of the nodes in meters in the y axis

Δz::Float64: The spacing of the nodes in meters in the z axis

Δt::Float64: The spacing of timesteps in seconds

Nₗ::Int64: The number of layers in the model

Δh::Float64: The layer height in nodes

ΔH::Int64: The layer height in meters

name::String: Just used for future reference of results

Boundary

HSSSimulations.Boundary.calcInds — Function

calcInds(res, ghost, ΔH, recoatLoadSet) -> Indices

Calculates the indices at the start of a load. For build load sets this includes 'imaginary' nodes that don't yet represent a volume with powder in it (but will after recoat). recoatLoadSet should be set to true if the current load set includes powder recoating, and false if it does not.

See Types.Indices for more details on the struct returned by this function.

HSSSimulations.Boundary.padWithGhost! — Function

padWithGhost!(
    pts::AbstractResult,
    cts::AbstractResult,
    ls,
    G
)

Wraps the temperature array with ghost cells and updates the Tᵗ⁻¹ array in G with the new value. The ghost cells are calculated based on a the boundaries for each face provided in the current load. See Types.Load for more details on these boundaries.