Internals

Documentation for any things not exported from modules.

Boundary

HSSSimulations.Boundary.updateInds! — Function

updateInds!(
    indStruct::HSSSimulations.Types.Indices,
    recoatDist,
    resSize,
    ghost
)

Updates an indices struct during a load step. Used to update the real and imaginary nodes during recoat of the powder layer based on the recoatDist (how far through the layer the powder has been deposited in number of nodes into the simulation area).

HSSSimulations.Boundary.ghostCalc! — Function

ghostCalc!(Tᵗ⁻¹, T, params, indices, gdist, κ)

Fills in the ghost nodes for the boundary on the face defined by indices using the boundaryHeatTransferRate function for the type of the boundary parametre (params) given.

HSSSimulations.Boundary.boundaryTemp — Function

boundaryTemp(ϕ⃗, T, κ, gdist) -> Any

Calculates a temperature for a ghost cell that will give the heat flux density (ϕ⃗, in wm⁻²) to a node with a temperature of T. gdist is the distance between the ghost and real node in meters, eg. for a z boundary it would be Δz.

HSSSimulations.Boundary.σ — Constant

Stefan-Boltzmann constant

Solver

HSSSimulations.Solver.fdmSolver! — Function

fdmSolver!(cts, indᵣ, p)

Uses the explicit time step finite difference heat form of the heat equation to calculate the new results using ghost cell boundary conditions.

In simple terms this takes the avarage for all nodes for the one around the one being calculated (from the previous time step), and uses the difference between that and the temperature of the current node (for the previous time step) multiplied by the fourier number Fo to calculate the temperature of that node for the current timestep. This is done for all nodes. As ghost nodes are used, this function is the same for all nodes and doesn't need special logic for boundary nodes.

HSSSimulations.Solver.timeSolver! — Function

timeSolver!(
    cts::AbstractResult,
    pts::AbstractResult,
    ls::HSSSimulations.Types.LoadStep,
    prob::Problem
)

Solves a timestep by calculating new material propeties, setting ghost node boundary conditions, using an explicit finited difference method and then conditionally recoating the poweder (if the load specifies). The results are stored in the Results struct cr that is passed into the function.

Arguments

cts : The struct that will be used to store the results from this solver.

pts : The results struct with the results from the previous step.

HSSSimulations.Solver.nanfiller! — Function

nanfiller!(cts, indᵢ)

Fills non real cells with NaNs instead of the undef value there currently. Using NaN instead of zero both allows for a check to see if the cell has been initialised and makes them not show up in plots.

This needs to be called on all loads (not just recoat loads) as on the loads before recoat there are still imaginary nodes that need filling (for pretty plotting).

HSSSimulations.Solver.loadSolver! — Function

loadSolver!(
    ls::HSSSimulations.Types.LoadStep,
    results,
    prob::Problem
) -> AbstractResult

Solves a single load and returns an array of the results.

HSSSimulations.Types.makeDescription — Function

makeDescription(problem::Problem) -> Any

Primaraly used for the show method for the Problem struct, but seperated into it's own function so the same formatting can be used for making a string to save to the results file. This is useful for having a summary of the simulation setup attached to the results for quick reference (The full problem struct is also saved, but that requires loading the results in a julia instance to read properly).

HSSSimulations.Solver.startMetadata — Function

startMetadata(problem, file)

Adds metadata to the results file. The data added here is stuff that is available at the start of the simulation, such as the problem description, the problem input and the start time.

HSSSimulations.Solver.finishMetadata — Function

finishMetadata(problem::Problem, file)

Adds metadata to the results file. The data added here is stuff that is not available until the end of the simulation, such as the melt max array, the results index array and the simulation end time.