Post Processing

HSSSimulations.PostProcessing — Module

Some functions to ease the post processing of the simulation results.

External

HSSSimulations.PostProcessing.makeIndex — Function

makeIndex(group)

Warning

This function is hard coded to only make an index for two levels deep, this is meant to work with the loadsets and loads as the two levels.

Internal