Solver

HSSSimulations.Solver — Module

The core logic for setting up and solving a simulation, including the heat transfer solver.

External

HSSSimulations.Solver.makeLogger — Function

makeLogger(debugGroups, logFile) -> LoggingExtras.TeeLogger

debugGroups is a list of groups of debug messages to log to the logFile. If it is set to true then all log groups in the package are used (see package_groups), if it is set to false then no log groups are used.

User log groups can be added to the list, simple add your own string to the list and then set debug messages with the _group field set to your string.

There is a special group called "misc" that will catch any group not in the package_groups list or in any additional group given by the user to the debugGroups list.

HSSSimulations.Solver.loadSetSolver! — Function

loadSetSolver!(
    loadSet<:AbstractLoadSet
    initResult::AbstractResult,
    layerNum::Int,
    prob::Problem{T,Gh,Mp,R,OR,B},
) where {T<:Any,Gh<:Any,Mp<:Any,R<:Any,OR<:Any,B<:Any}

This function is run once for each LoadSet passed in to the problem, and dispatched on the type of that load set. For basic load sets, like FixedLoadSet, it should basically just pass the arguments on to the innerLoadSetSolver! function, the one exception being the list of loads being passed in instead of the AbstractLoadSet struct. For more complicated load sets, like LayerLoadSet, this should include the logic used for that load set. For LayerLoadSet this is a loop that loops over the list of loads once for each layer, incrementing the layer number as it goes.

For examples see the source code for the two default implementations.

HSSSimulations.Types.AbstractLoadSet — Type

abstract type AbstractLoadSet <: HSSSimulations.Types.AbstractSimProperty

This struct holds the information for a load set, primarily the list of loads, but also some extra bits. Subtypes should implement a method for the Solver.loadSetSolver! function.

HSSSimulations.Solver.loadSetSolver! — Method

loadSetSolver!(
    loadSet::FixedLoadSet,
    initResult::AbstractResult,
    layerNum::Int64,
    prob::Problem{T, Gh, Mp, R, OR, B}
) -> Tuple{AbstractResult, Int64}

HSSSimulations.Solver.loadSetSolver! — Method

loadSetSolver!(
    loadSet::LayerLoadSet,
    initResult::AbstractResult,
    initLayerNum::Int64,
    prob::Problem{T, Gh, Mp, R, OR, B}
) -> Tuple{AbstractResult, Int64}

Internal

HSSSimulations.Solver.innerLoadSetSolver! — Function

innerLoadSetSolver!(
    loads::Vector{Load},
    initResult::AbstractResult,
    layerNum,
    prob::Problem;
    name,
    recoat,
    prevLayerNum
)

Solves a load set of a problem, iterating over and solving the loads in the given load set in order. The results for a load is saved to the file after solving it.

If the list of loads passed given include powder deposition, the recoat kwarg should be set to true, with the layerNum set to the thickness of the powder bed in layers that is desired at the end of the loads. If this is more than one layer thicker than the previous value, then the previous values should also be passed in to prevLayerNum.

It also helps to provide a unique name. This is the name that will be used to store the results from runnning the list of loads in the output file, and also will be used as the label in the progress meter. If no name is given then the current time is used instead. If the name given is not unique then you will get a run time error when it tries to save results to a place that is already taken.

HSSSimulations.Solver.fdmSolver! — Function

fdmSolver!(cts, indᵣ, p)

Uses the explicit time step finite difference heat form of the heat equation to calculate the new results using ghost cell boundary conditions.

In simple terms this takes the avarage for all nodes for the one around the one being calculated (from the previous time step), and uses the difference between that and the temperature of the current node (for the previous time step) multiplied by the fourier number Fo to calculate the temperature of that node for the current timestep. This is done for all nodes. As ghost nodes are used, this function is the same for all nodes and doesn't need special logic for boundary nodes.

HSSSimulations.Solver.timeSolver! — Function

timeSolver!(
    cts::AbstractResult,
    pts::AbstractResult,
    ls::HSSSimulations.Types.LoadStep,
    prob::Problem
)

Solves a timestep by calculating new material propeties, setting ghost node boundary conditions, using an explicit finited difference method and then conditionally recoating the poweder (if the load specifies). The results are stored in the Results struct cr that is passed into the function.

Arguments

cts : The struct that will be used to store the results from this solver.

pts : The results struct with the results from the previous step.

HSSSimulations.Solver.nanfiller! — Function

nanfiller!(cts, indᵢ)

Fills non real cells with NaNs instead of the undef value there currently. Using NaN instead of zero both allows for a check to see if the cell has been initialised and makes them not show up in plots.

This needs to be called on all loads (not just recoat loads) as on the loads before recoat there are still imaginary nodes that need filling (for pretty plotting).

HSSSimulations.Solver.loadSolver! — Function

loadSolver!(
    ls::HSSSimulations.Types.LoadStep,
    results,
    prob::Problem
) -> AbstractResult

Solves a single load and returns an array of the results.

HSSSimulations.Types.makeDescription — Function

makeDescription(problem::Problem) -> Any

Primaraly used for the show method for the Problem struct, but seperated into it's own function so the same formatting can be used for making a string to save to the results file. This is useful for having a summary of the simulation setup attached to the results for quick reference (The full problem struct is also saved, but that requires loading the results in a julia instance to read properly).

HSSSimulations.Solver.startMetadata — Function

startMetadata(problem, file)

Adds metadata to the results file. The data added here is stuff that is available at the start of the simulation, such as the problem description, the problem input and the start time.

HSSSimulations.Solver.finishMetadata — Function

finishMetadata(problem::Problem, file)

Adds metadata to the results file. The data added here is stuff that is not available until the end of the simulation, such as the melt max array, the results index array and the simulation end time.