API

Solver

HSSSimulations.Solver.problemSolver — Function

problemSolver(problem::Problem) -> Tuple{Any, Any}

The main function that is called to solve a simulation and the struct that defines the problem to simulate.

Takes a fully defined problem and solves it, saving the solution to disk and returning the filename where it is saved. See Problem for how to define the problem.

HSSSimulations.Types.Problem — Method

Problem(
;
    geometry,
    matProp,
    params,
    loadSets,
    init,
    file,
    initLay,
    ink,
    description,
    otherResults,
    options
)

Assemble a problem out of its components.

Arguments

`geometry::Geometry`: [`Geometry`](@ref)
`matProp::AbstractMatProp`: [`AbstractMatProp`](@ref)
`params::AbstractProblemParams`: [`AbstractProblemParams`](@ref)
`loadSets::Vector{AbstractLoadSet}`: [`AbstractLoadSet`](@ref)
`ink::Ink`: [`Ink`](@ref)
`init::AbstractResult`: [`AbstractResult`](@ref)
`otherResults::AbstractOtherResults=OtherResults()`: [`AbstractOtherResults`](@ref)
`initLay::Int`: The thickness of powder to use for preheat loads, given in number of layers thick
`file::String`: File name
`description::String=""`: A short description of what is being simulated
`options::Options=Options()`: Simulation options

Simulation Geometry

HSSSimulations.Types.Geometry — Method

Geometry(
simSize, Δx, Δt;
Δy=Δx, Δz=Δx, name="NA", Δh=0,
offset=(0.0, 0.0), buildSize=nothing,
force=false,

)

Constructor for the Geometry type that is is used to store all of the geometry information (and time step length for some reason) for a rectangular build volume of the machine being simulated (given as the buildSize). It also saves the information for the subset of the build volume to actually be simulated (of size simSize, offset form the machine origin by offset), if the full build volume is not being simulated. If no buildSize is given then it is assumed to be just big enougth to fit the simSize with the given offset.

Δh is the layer height in meters. If it is given as 0 (or not given) then it is assumed that the simulation isn't representing a full build, but instead something like the preheat or cooldown phase. In this case no layer recoat logic can be run (make sure not to include a recoating Types.Load).

Δt is the time step (in seconds) and Δx, Δy and Δz are the node spacing (in meters). If not given then Δy and Δz default to the same as Δx. The timestep is included in the geometry as it is tied to the node spacing when it comes to making a stable simulation for the explicit finite difference method used in this model.

If the force argement is given then the divisible errors will be suppressed, this will result in the geometry not being properly represented.

Use With Caution

Use the force argument with great caution. It was only added to allow for the creation of geometries that were blocked due to floating point math errors. If it is used when things aren't actually divisible then it will result in the geometry not being properly represented, and a disconnect between what you think you are simulating and what is actually being simulated.

Ink

HSSSimulations.Types.Ink — Type

struct Ink{T}

Defines the volume of the ink placement within that (and therefore hopefully the part to be made).

Fields

nodes::Array{T, 3} where T: The emmisivity of the models nodes, set to eₚ for nodes without ink.

name::String: Just used for future reference of results

This is the emmisivity relative to the lamp. So the emmisivity of the ink over the range of the wavelengths that the lamp outputs, scaled by the relative output power of the lamp at those wavelengths.

Materials

HSSSimulations.Material.PA2200 — Function

PA2200(
    geometry::Geometry
) -> MatProp{Interpolations.Extrapolation{Float64, 2, Interpolations.ScaledInterpolation{Float64, 2, Interpolations.BSplineInterpolation{Float64, 2, Matrix{Float64}, Interpolations.BSpline{Interpolations.Linear{Interpolations.Throw{Interpolations.OnGrid}}}, Tuple{Base.OneTo{Int64}, Base.OneTo{Int64}}}, Interpolations.BSpline{Interpolations.Linear{Interpolations.Throw{Interpolations.OnGrid}}}, Tuple{UnitRange{Int64}, UnitRange{Int64}}}, Interpolations.BSpline{Interpolations.Linear{Interpolations.Throw{Interpolations.OnGrid}}}, Interpolations.Flat{Nothing}}, _A, _B, _C, _D, _E, _F, typeof(PA_Ċ), Array{Float64, 3}} where {_A, _B, _C, _D, _E, _F}

An example material based on PA2200, using a rate of consolidation based on melt state. With eyeball correction to consolidation rate.

Results

HSSSimulations.Res.Result — Type

struct Result{P<:AbstractArray} <: AbstractResult

The results from a single timestep, use directly to create the initial conditions. Also created for each time step during the simulation.

Fields

T::AbstractArray: Temperature

M::AbstractArray: Melt state

C::AbstractArray: Consolidation state

t::Float64: Time of timestep

tₚ::Float64: The progress through the load step (0=start, 1=end)

HSSSimulations.Res.Result — Method

Result(geomSize, Tᵢ, Mᵢ, Cᵢ, t, tₚ) -> Result

Create a result with uniform fields

HSSSimulations.Res.Result — Method

Result(T, M, C, t, tₚ) -> Result

Re-expose the default constructor

HSSSimulations.Res.Result — Method

Result(geomSize, t) -> Result

Create a result filled with zeros

HSSSimulations.Res.Result — Method

Result(geomSize, t, tₚ) -> Result

Create an empty result

Boundaries

HSSSimulations.Boundary.basicLoad — Function

basicLoad(
    tₗ,
    skip
) -> Load{SymetryBoundary, SymetryBoundary, SymetryBoundary, SymetryBoundary}

A basic Types.Load with a conduction boundary on the bottom surface and a convection boundary on the top. All other surfaces are symetrical boundaries.

Examples

julia> loadStep = basicLoad(5, 2)
  x₁ : SymetryBoundary
  x₂ : SymetryBoundary
  y₁ : SymetryBoundary
  y₂ : SymetryBoundary
  z₁ : ConductionBoundary
  z₂ : ConvectionBoundary
  name : NA
  tₗ : 5.0
  skip : 2

This returns a single Load, for a single load step. To make a load set you will need an array of Loads.

HSSSimulations.Boundary.BasicProblemParams — Type

A basic implementation of a Types.AbstractProblemParams struct to go along with basicLoad. For a more elaborate example see HSSParams

Fields

condCoef::Float64: The contact conduction coefficent for the bottom face

condTemp::Any: The temperature of the surface in contact with the bottom face

convCoef::Float64: The convection coefficent for the top face to the air above

convTemp::Any: The temperature of the air in contact with the top face

Settings

HSSSimulations.Types.Options — Type

struct Options

Fields

compress::Union{Bool, TranscodingStreams.Codec}: How to compress the results file, can be set to true (to compress), false (to leave uncompressed) or to a specific compression algorithm (see the JLD2 documentation for more details)

debug::Union{Bool, Vector{String}}: Whether or not to log debug information, can accept a list of strings to select only some debugging groups, see package_groups

showProgress::Union{Bool, Float64}: Whether or not to show the progres meter, if a number is given that is used as the update interval

notify::Bool: If true, simulations finishing will send a system notification using Alert.jl

The debug option is passed to the logGroups option of Solver.makeLogger, check that out for more information and package_groups for what log groups are available by default.

Note

Depending on settings, the debug option might log a lot of things, the log file could end up somewhere in the region of 4x the size of the compressed results file, so make sure you clean them up after you're done.

HSSSimulations.Types.package_groups — Constant

A list of all of the log groups used in this package. They log the following things:

"core": the start of a problem, loadstep, load or timestep has started

"solver": the fdm solver

"mat": material model

"bound": boundary condition

"b_adv": recoating and moving object boundaries

"hss": HSS example functions

Results File Structure

The last part of the API to cover is the format of the simulation results that are saved. These use the JLD2 package to save to a Hierarchical Data Format version 5 (HDF5) based format. Most of the information stored should be readable by any HDF5 compatible software or libraries, except the input problem, which requires the software to understand julia types (the easiest way is to just use juila, it's mostly saved in case it needs to be rerun).

By default, this is stored compressed using the ZlibCompressor compressor.

The results have the following structure:

Tree                           Description
Results
│
├─ Description                 - An overview of the problem that has been solved
├─ Input                       - The full problem struct that has been solved
├─ Start_Time                  - The computer's clock time at the end of the simulation
├── Results
│   ├─ MeltMax                 - The maximum melt state reached in the simulation
│   │
│   ├── Preheat-1              - List of loads run during the preheating load set
│   │   ├─ 1
│   │   ├─ 2
│   │   └─ ⋯ (3 more entries)
│   │
│   ├── Layer-2                - 1st Layer
│   │   ├─ 1                   - 1st Layer's 1st Load
│   │   ├─ 2                   - 1st Layer's 2nd Load
│   │   └─ ⋯ (6 more entries)  - And so on for the remaining loads
│   │
│   ├── Layer-3                - 2nd Layer
│   │   ├─ 1 (5 entries)
│   │   ├─ 2
│   │   └─ ⋯ (6 more entries)
│   │
│   ├─ ⋯ (25 more entries)     - And so on for the remaining layers
│   │
│   └── Cooldown-28            - List of loads run during the cooldown load set
│       ├─ 1
│       ├─ 2
│       └─ ⋯ (3 more entries)
│
├─ Results_Index               - A list of all load results indices within this file
├─ Finish_Time                 - The computer's clock time at the end of the simulation
└─ _types                      - Ignore (Used internally by JLD2)

Where each of the loads has these fields:

Tree           Description

Load
├─ name        - The load's name
├─ time        - 1D array of the times of the load's time steps
├─ T           - 4D array of temperatures (X, Y, Z, time step)
├─ M           - 4D array of melt states (X, Y, Z, time step)
└─ C           - 4D array of consolidation states (X, Y, Z, time step)

Tip

Res.loadStepSaver, Solver.otherResults, are the two functions used for saving simulation results to the file. so looking at their implementation might help with if anything is not covered here. Solver.startMetadata and Solver.finishMetadata are also used to save a few extra bits of metadata to the file. I'd also recommend a tool like HDFView to get an idea for the structure of the results.